subProj 4 CAPL Converter/requirements.txt
pytest==8.0.0
pytest-html==4.1.1
python-can==4.3.1
